requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
opencv-python-headless==4.8.0.76
numpy==1.24.4
scikit-learn==1.3.0
matplotlib==3.7.2
pillow==10.0.0
requests==2.31.0